now able to pull data from parse. go to doctors fragment to see changes and how i did this, also there is a link to a video.
go to the new java class i made called DoctorInfo.java to see how to create the class that holds the information we get from parse.

Minimum changes done to DoctorAdapter other than changing the type of list i was passing it

I had to add the ParseApplication.java because parse would not work for me and give me errors if i didn't add it. I do not know if ya'll will need it or what it would do to your program.