Can now set swipe listener to items, swipe right on a doctor item and popup dialog button will appear asking if you wish to delete it. If yes object is deleted from parse and list is updated. I think i did that in doctor adapter.java. Needs the new OnSwipeTouch Listener.java to function. Don't worry about that class its generic. What happens when i swipe is defined in the adapter.
Also added a add doctor floating button to the doctors fragment.

Added stuff to the MainActivity so that when you select back it goes back to the last fragment. however there are still some bugs were layouts overlay past layouts. I'm trying to fix it but i'll do it later.