Fixed doctor bugs.
Swipe to delete is very hard to do on the cell phone
not sure why this is, but it takes me about 10 swipes before it works
FIXED: made tabs match parent to take entire width of screen and made them take more space on the top and bottom

Viewing a specific doctor and pressing the call button crashes the app
UNKNOWN: doesn’t crash on my phone or in VM

Create a new doctor, then try to view that new doctor, app crashes
SEMI-FIXED commented out setting the new date in parse. i haven’t figured out who I am going to fix it but for now it doesn’t save user input on date to parse. but for now app will not crash