Vitals page works. however it is in a new activty for now, i will see about how to put it in a fragment soon.
Fixed it to where fragments do not pile on top of each other when you go to a certain  page and back out. Go to the doctor activity page and edit page to see where i added code (i think thats where the code is). Just had to add pop off back stack code i think. Should be where the code for the onClick is at.